+++ tests/test_m4i.py
@@ -1,5 +1,6 @@
 import unittest
 
+import matplotlib.pyplot as plt
 import pydantic
 
 from ontolutils.ex.m4i import TextVariable, NumericalVariable, Tool, ProcessingStep
@@ -68,3 +69,76 @@ class TestM4i(unittest.TestCase):
         self.assertEqual(numerical_variable2.hasNumericalValue, 1.0)
         self.assertEqual(numerical_variable2.hasMaximumValue, 2.0)
         self.assertEqual(numerical_variable2.hasVariableDescription, 'Variable description')
+
+    def test_to_pint(self):
+        numerical_variable = NumericalVariable(
+            hasUnit='mm/s',
+            hasNumericalValue=1.0,
+            hasMaximumValue=2.0,
+            hasVariableDescription='Variable description')
+        pint_quantity = numerical_variable.to_pint()
+        self.assertEqual(pint_quantity.magnitude, 1.0)
+        self.assertEqual(str(pint_quantity.units), 'millimeter / second')
+
+        numerical_variable_array = NumericalVariable(
+            hasUnit='mm/s',
+            hasNumericalValue=[1.0, 2.0, 3.0],
+            hasMaximumValue=5.0,
+            hasVariableDescription='Variable description')
+        pint_quantity_array = numerical_variable_array.to_pint()
+        self.assertEqual(pint_quantity_array.magnitude.tolist(), [1.0, 2.0, 3.0])
+        self.assertEqual(str(pint_quantity_array.units), 'millimeter / second')
+
+        numerical_variable_from_pint = NumericalVariable.from_pint(pint_quantity)
+        self.assertEqual(
+            numerical_variable_from_pint.hasUnit,
+            'http://qudt.org/vocab/unit/MilliM-PER-SEC'
+        )
+        self.assertEqual(numerical_variable_from_pint.hasNumericalValue, 1.0)
+
+        numerical_variable_from_pint = NumericalVariable.from_pint(
+            pint_quantity_array,
+            id="http://example.org/variable/vfr",
+            hasSymbol='vfr'
+        )
+        self.assertEqual(
+            numerical_variable_from_pint.hasUnit,
+            'http://qudt.org/vocab/unit/MilliM-PER-SEC'
+        )
+        self.assertEqual(numerical_variable_from_pint.hasNumericalValue, [1.0, 2.0, 3.0])
+        self.assertEqual(numerical_variable_from_pint.hasSymbol, 'vfr')
+        ttl = numerical_variable_from_pint.serialize("ttl")
+        self.assertEqual(ttl, """@prefix m4i: <http://w3id.org/nfdi4ing/metadata4ing#> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+
+<http://example.org/variable/vfr> a m4i:NumericalVariable ;
+    m4i:hasNumericalValue 1e+00,
+        2e+00,
+        3e+00 ;
+    m4i:hasSymbol "vfr" ;
+    m4i:hasUnit <http://qudt.org/vocab/unit/MilliM-PER-SEC> .
+
+""")
+
+    def test_to_xarray(self):
+        numerical_variable = NumericalVariable(
+            id='http://example.org/variable/vfr',
+            label=["Volume Flow Rate@en", "Volumenstrom@de", "Débit volumique"],
+            hasUnit='mm/s',
+            hasNumericalValue=[1.0, 2.0, 3.0],
+            hasVariableDescription='Variable description',
+            hasSymbol='vfr'
+        )
+        ttl_orig = numerical_variable.serialize("ttl")
+        xarray_dataarray = numerical_variable.to_xarray(language="de")
+        self.assertEqual(xarray_dataarray.has_symbol, 'vfr')
+        self.assertEqual(xarray_dataarray.attrs['has_variable_description'], 'Variable description')
+        self.assertEqual(xarray_dataarray.attrs['units'], 'mm/s')
+        self.assertEqual(xarray_dataarray.values.tolist(), [1.0, 2.0, 3.0])
+
+        # convert back to NumericalVariable
+        numerical_variable_from_xarray = NumericalVariable.from_xarray(
+            xarray_dataarray
+        )
+        ttl_after_conversions = numerical_variable_from_xarray.serialize("ttl")
+        self.assertEqual(ttl_orig, ttl_after_conversions)